Survey
A survey application for a school project. Contains a Survey class that
holds 10 questions and 10 responses to each question, as well as methods
for displaying survey results. Also has a Survey Test class for testing
the Survey class during development, and a final Survey Conductor class
that actually conducts the survey after receiving questions from the
user.